aggiunta la possibilita' di riprendere il download anche avendolo interrotto

+++ Dataset/imagedownloader.py
@@ -25,15 +25,16 @@ class ThreadedDownload(threading.Thread):
 
     def run(self):
         failed = 0
-        for line in self.url_list:
+        for i, line in enumerate(self.url_list):
+            if i % 100 == 0:
+                print('Thread {n} processed {i} urls'.format(n=self.name, i=i))
             url = line.split('\t')[1]
             url = url.replace('\n', '')
             url = url.strip()
             try:
                 ImageNetDownloader.download_file(url, self.dir_path)
-
             except Exception:
-                print('Fail to download : ' + url)
+                # print('Fail to download : ' + url)
                 failed += 1
         print("Thread: {name} \nDownloaded images: {images} \nFailed: {failed}".format(
             name=self.name,
@@ -60,30 +61,30 @@ class ImageNetDownloader:
 
         if desc:
             filename = os.path.join(desc, filename)
+        if not os.path.isfile(filename):
+            with open(filename, 'wb') as f:
+                meta = u.info()
+                meta_func = meta.getheaders if hasattr(meta, 'getheaders') else meta.get_all
+                meta_length = meta_func("Content-Length")
+                file_size = None
+                if meta_length:
+                    file_size = int(meta_length[0])
+                # print("Downloading: {0} Bytes: {1}".format(url, file_size))
 
-        with open(filename, 'wb') as f:
-            meta = u.info()
-            meta_func = meta.getheaders if hasattr(meta, 'getheaders') else meta.get_all
-            meta_length = meta_func("Content-Length")
-            file_size = None
-            if meta_length:
-                file_size = int(meta_length[0])
-            # print("Downloading: {0} Bytes: {1}".format(url, file_size))
+                file_size_dl = 0
+                block_sz = 8192
+                while True:
+                    buffer = u.read(block_sz)
+                    if not buffer:
+                        break
 
-            file_size_dl = 0
-            block_sz = 8192
-            while True:
-                buffer = u.read(block_sz)
-                if not buffer:
-                    break
+                    file_size_dl += len(buffer)
+                    f.write(buffer)
 
-                file_size_dl += len(buffer)
-                f.write(buffer)
-
-                status = "{0:16}".format(file_size_dl)
-                if file_size:
-                    status += "   [{0:6.2f}%]".format(file_size_dl * 100 / file_size)
-                status += chr(13)
+                    status = "{0:16}".format(file_size_dl)
+                    if file_size:
+                        status += "   [{0:6.2f}%]".format(file_size_dl * 100 / file_size)
+                    status += chr(13)
 
         return filename
 
@@ -105,7 +106,7 @@ if __name__ == '__main__':
     # count = 0
     new_urls = 'new_links.txt'
     dest_path = 'resources'
-    n_threads = 32
+    n_threads = 60
     tot_images = int(2e5)
     random.seed(3)  # force shuffle order
     if not os.path.isfile(new_urls):
